techpack: camera: Import minimal xiaomi camera modifications
From: psyche-r-oss

+++ techpack/camera/drivers/cam_sensor_module/cam_sensor_io/cam_sensor_io.h
@@ -106,7 +106,7 @@ int32_t camera_io_dev_erase(struct camera_io_master *io_master_info,
  * This API abstracts poll functionality based on master type
  */
 int32_t camera_io_dev_poll(struct camera_io_master *io_master_info,
-	uint32_t addr, uint16_t data, uint32_t data_mask,
+	uint32_t addr, uint32_t data, uint32_t data_mask,
 	enum camera_sensor_i2c_type addr_type,
 	enum camera_sensor_i2c_type data_type,
 	uint32_t delay_ms);
